server/requirements.txt
flask
flask-cors
beautifulsoup4
requests
google-generativeai
sentence-transformers
pinecone-client
python-dotenv
waitress
gunicorn